backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.3.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
transformers>=4.30.0
torch
openai
joblib
